split header and sidebar nav
