--- src/components/Header.astro
@@ -1,34 +1,0 @@
----
-import { GET as getReviewCount } from "src/pages/api/sr/review/count";
-import { Anchor } from "ui/anchor";
-import Logo from "./Logo.astro";
-import { HamburgerButton } from "./solid/HamburgerButton";
-import { Nav } from "./solid/Nav";
-import { isUserAdmin } from "libs/auth/admin";
-import Notifications from "./Notifications.astro";
-
-const user = Astro.locals.user;
-const reviewCount =
-  user && ((await (await getReviewCount(Astro)).json()) as number);
----
-
-<header class="bg-primary p-4 sticky top-0 z-50 min-h-[76px]">
-  <div class="lg:container mx-auto flex justify-between items-center">
-    <HamburgerButton client:load />
-    <Anchor class="text-3xl font-bold text-white" href="/"><Logo /></Anchor>
-    <div class="justify-between gap-5 hidden lg:flex">
-      <Nav
-        client:load
-        loggedIn={!!user}
-        reviewCount={reviewCount}
-        isAdmin={isUserAdmin(user ?? undefined)}
-      >
-        <Notifications loggedIn={!!user} />
-      </Nav>
-    </div>
-
-    <div class="flex lg:hidden">
-      <Notifications loggedIn={!!user} />
-    </div>
-  </div>
-</header>
