+++ src/components/header/Header.astro
@@ -1,0 +1,18 @@
+---
+import { Anchor } from "ui/anchor";
+import Logo from "../Logo.astro";
+import { HamburgerButton } from "../solid/HamburgerButton";
+import HeaderNav from "./HeaderNav.astro";
+
+const user = Astro.locals.user;
+---
+
+<header class="bg-primary p-4 sticky top-0 z-50 min-h-[76px]">
+  <div class="lg:container mx-auto flex justify-between items-center">
+    <HamburgerButton client:load />
+    <Anchor class="text-3xl font-bold text-white" href="/"><Logo /></Anchor>
+    <div class="justify-between gap-5 hidden lg:flex">
+      <HeaderNav />
+    </div>
+  </div>
+</header>
