+++ src/layouts/Layout.astro
@@ -5,7 +5,7 @@ import "../styles/global.css";
 
 import { Sidebar } from "@components/solid/Sidebar/Sidebar";
 import { Toaster } from "solid-toast";
-import Header from "../components/Header.astro";
+import Header from "../components/header/Header.astro";
 import { Image } from "astro:assets";
 import github from "public/github-mark.svg";
 import telegram from "public/telegram.svg";
